fix decimation % logic

+++ Libs/Optimize/ParticleSystem/VtkMeshWrapper.cpp
@@ -809,7 +809,7 @@ void VtkMeshWrapper::GetIGLMesh(Eigen::MatrixXd& V, Eigen::MatrixXi& F) const
 vtkSmartPointer<vtkPolyData> VtkMeshWrapper::Decimated(unsigned long target_tris) const
 {
   const auto total_tris = this->triangles_.size();
-  const double target_reduction = std::min(1.0, static_cast<double>(total_tris - target_tris) / total_tris);
+  const double target_reduction = std::min(1.0, ((double)total_tris - (double)target_tris) / (double)total_tris);
   if(target_reduction <= 0.0) {
     //todo in the event the user doesn't require decimation, we unnecessarily make a copy of the mesh.
     auto decimated = vtkSmartPointer<vtkPolyData>::New();
